--- src/main/java/com/medman/repositories/MedicationRepository.java
@@ -1,8 +1,0 @@
-package com.medman.repositories;
-
-import com.medman.models.Medication;
-import org.springframework.data.repository.CrudRepository;
-
-public interface MedicationRepository extends CrudRepository<Medication, Long> {
-
-}
