--- src/main/java/com/medman/repositories/PrescriptionRepository.java
@@ -1,8 +1,0 @@
-package com.medman.repositories;
-
-import com.medman.models.Prescription;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.repository.CrudRepository;
-
-public interface PrescriptionRepository extends CrudRepository<Prescription, Long> {
-}
