--- src/main/java/com/medman/repositories/UserRepository.java
@@ -1,13 +1,0 @@
-package com.medman.repositories;
-
-import com.medman.models.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-
-    User findByUsernameIgnoreCase(String username);
-
-    User findByEmailIgnoreCase(String email);
-
-    User findByUsernameOrEmail(String username, String email);
-}
