+++ src/main/java/com/medman/services/AuthException.java
@@ -1,8 +1,8 @@
-package com.medman.services;
-
-public class AuthException extends Exception {
-
-    public AuthException(String message) {
-        super(message);
-    }
-}
+//package com.medman.services;
+//
+//public class AuthException extends Exception {
+//
+//    public AuthException(String message) {
+//        super(message);
+//    }
+//}
