+++ src/main/java/com/medman/services/AvatarService.java
@@ -1,10 +1,10 @@
-package com.medman.services;
-
-import org.springframework.web.multipart.MultipartFile;
-
-import java.io.IOException;
-
-public interface AvatarService {
-
-    UploadedAvatarInfo upload(MultipartFile file) throws IOException, UnsupportedFormatException;
-}
+//package com.medman.services;
+//
+//import org.springframework.web.multipart.MultipartFile;
+//
+//import java.io.IOException;
+//
+//public interface AvatarService {
+//
+//    UploadedAvatarInfo upload(MultipartFile file) throws IOException, UnsupportedFormatException;
+//}
