+++ src/main/java/com/medman/services/FileNameGenerator.java
@@ -1,6 +1,6 @@
-package com.medman.services;
-
-public interface FileNameGenerator {
-
-    String getFileName(String filename, String prefix);
-}
+//package com.medman.services;
+//
+//public interface FileNameGenerator {
+//
+//    String getFileName(String filename, String prefix);
+//}
