+++ src/main/java/com/medman/services/FileNameGeneratorImpl.java
@@ -1,11 +1,11 @@
-package com.medman.services;
-
-import org.springframework.stereotype.Service;
-
-@Service("fileNameGenerator")
-public class FileNameGeneratorImpl implements FileNameGenerator {
-    @Override
-    public String getFileName(String filename, String prefix) {
-        return prefix + filename.hashCode() + System.currentTimeMillis();
-    }
-}
+//package com.medman.services;
+//
+//import org.springframework.stereotype.Service;
+//
+//@Service("fileNameGenerator")
+//public class FileNameGeneratorImpl implements FileNameGenerator {
+//    @Override
+//    public String getFileName(String filename, String prefix) {
+//        return prefix + filename.hashCode() + System.currentTimeMillis();
+//    }
+//}
