+++ src/main/java/com/medman/services/UnsupportedFormatException.java
@@ -1,9 +1,9 @@
-package com.medman.services;
-
-public class UnsupportedFormatException extends Exception {
-
-    public UnsupportedFormatException(String message) {
-        super(message);
-    }
-}
-
+//package com.medman.services;
+//
+//public class UnsupportedFormatException extends Exception {
+//
+//    public UnsupportedFormatException(String message) {
+//        super(message);
+//    }
+//}
+//
